Completed video playback * Renamed AmazonPlayer.py into playback.py * Moved additional playback related functions into playback.py * Converted most of playback related functions into private * Renamed `getStreams` into `ParseStream` to better reflect its usage * Moved `sleep` and `jsonRPC` into common.py * Moved age restriction code into a separate ages.py module * Converted the set of age globals into dynamic getters
